+++ packages/analytics/src/index.ts
@@ -26,3 +26,4 @@ export * from './marketBeta';
 export * from './seasonality';
 export * from './valuationHistory';
 export * from './fundingAnalytics';
+export * from './bookAnalytics';